+++ src/item/Hoe.cpp
@@ -10,7 +10,7 @@ namespace Game3 {
 	Hoe::Hoe(ItemID id_, std::string name_, MoneyCount base_price, Durability max_durability):
 		Tool(id_, std::move(name_), base_price, 0.f, max_durability, "base:attribute/hoe"_id) {}
 
-	bool Hoe::use(Slot slot, ItemStack &stack, const Place &place, Modifiers, std::pair<double, double>) {
+	bool Hoe::use(Slot slot, ItemStack &stack, const Place &place, Modifiers, std::pair<float, float>) {
 		Realm &realm = *place.realm;
 		Tileset &tileset = realm.getTileset();
